axiom-fastapi/pyproject.toml
[project]
name = "axiom-fastapi"
version = "0.1.0"
description = "Axiom FastAPI integration — base package for FastAPI applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "axiom-core",
    "fastapi>=0.115.0,<1",
    "uvicorn>=0.29.0,<1",
    "tomlkit>=0.12,<1",
]

[project.optional-dependencies]
gunicorn = ["gunicorn>=21.0", "uvicorn[standard]>=0.29"]
rate-limiter = ["limits>=3.0,<4", "axiom-redis", "pydantic-settings>=2.0,<3"]

[dependency-groups]
dev = [
    "mypy>=1.16.0,<2",
    "pytest>=8.0.0,<9",
    "pytest-asyncio>=0.24.0,<1",
    "httpx>=0.27.0,<1",
    "ruff>=0.11.13,<1",
    "fakeredis>=2.0,<3",
    "limits>=3.0,<4",
    "axiom-redis",
    "pydantic-settings>=2.0,<3",
    "testcontainers[redis]>=4.0,<5",
]

[tool.uv.sources]
axiom-core = { workspace = true }
axiom-redis = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/axiom"]

[tool.mypy]

[[tool.mypy.overrides]]
module = "uvloop.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "gunicorn.*"
ignore_missing_imports = true
